カメラ、回転、平行移動、ズーム追加

+++ SysMouse.h
@@ -36,7 +36,7 @@ public:
 //	int	mx;	//	マウスカーソル移動量
 //	int	my;
 
-	int	wheel;
+	double	wheel;
 /*
 	struct 
 	{
